% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riceidconverter.R
\name{riceidconverter}
\alias{riceidconverter}
\title{Convert Biological ID from One to Another for Oryza Sativa}
\description{
\code{RiceIDConverter} Convert one biological ID to another for Oryza sativa,
such as MSU to SYMBOL.
}
\author{
Xiang LI <ynaulx@gmail.com>
}
